Updated database and api method for pagination.

+++ src/main/java/org/example/webapitableau/models/RecipeEntity.java
@@ -28,8 +28,8 @@ public class RecipeEntity {
     }
 
     // Getters and Setters
-    public Long getId() {
-        return id;
+    public Integer getId() {
+        return Math.toIntExact(id);
     }
 
     public void setId(Long id) {
